Move cdp code to webdriver.js
